requirements.txt
fastapi==0.109.2
psycopg2==2.9.9
python-dotenv
